# docs/swagger.yaml
definitions:
  models.AttendanceCreate:
    properties:
      amount:
        example: 1234.56
        type: number
      date:
        example: "2022-01-01"
        type: string
      employee_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      hours:
        maximum: 24
        type: integer
      is_holiday:
        default: false
        example: false
        type: boolean
      role:
        enum:
        - presente
        - tarde
        - parcial
        - ausente
        type: string
    required:
    - amount
    - date
    - employee_id
    type: object
  models.AttendanceLaundry:
    properties:
      amount:
        type: number
      created_at:
        type: string
      date:
        type: string
      employee:
        $ref: '#/definitions/models.EmployeeLaundry'
      employee_id:
        type: string
      hours:
        maximum: 24
        type: integer
      id:
        type: string
      is_holiday:
        type: boolean
      role:
        enum:
        - presente
        - tarde
        - parcial
        - ausente
        type: string
      updated_at:
        type: string
    type: object
  models.AttendanceUpdate:
    properties:
      amount:
        example: 1234.56
        type: number
      date:
        example: "2022-01-01"
        type: string
      employee_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      hours:
        maximum: 24
        type: integer
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      is_holiday:
        default: false
        example: false
        type: boolean
      role:
        enum:
        - presente
        - tarde
        - parcial
        - ausente
        type: string
    required:
    - amount
    - date
    - employee_id
    - id
    type: object
...
  models.DateBetween:
    properties:
      date_from:
        example: "2022-01-01"
        type: string
      date_to:
        example: "2022-01-01"
        type: string
    required:
    - date_from
    - date_to
    type: object
...
  models.EmployeeLaundry:
    properties:
      address:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
...
  models.Response:
    properties:
      body: {}
      message:
        type: string
      status:
        type: boolean
    type: object
...
info:
  contact: {}
  description: This is a api to app gestioncar
  termsOfService: http://swagger.io/terms/
  title: APP GESTIONCAR
  version: "1.0"
paths:
  /attendance/{id}:
    get:
      consumes:
      - application/json
      description: Get Attendance by ID
      parameters:
      - description: Workplace Token
        in: header
        name: X-Workplace-Token
        required: true
        type: string
      - description: ID of Attendance
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                body:
                  $ref: '#/definitions/models.AttendanceLaundry'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Get Attendance By ID
      tags:
      - Attendance
  /attendance/create:
    post:
      consumes:
      - application/json
      description: Create Attendance by given workplace
      parameters:
      - description: Workplace Token
        in: header
        name: X-Workplace-Token
        required: true
        type: string
      - description: Employee body
        in: body
        name: attendanceCreate
        required: true
        schema:
          $ref: '#/definitions/models.AttendanceCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Create Attendance
      tags:
      - Attendance
  /attendance/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Attendance by ID
      parameters:
      - description: Workplace Token
        in: header
        name: X-Workplace-Token
        required: true
        type: string
      - description: ID of Attendance
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Delete Attendance
      tags:
      - Attendance
  /attendance/get_all:
    get:
      consumes:
      - application/json
      description: Get all attendances by workplace required auth token
      parameters:
      - description: Workplace Token
        in: header
        name: X-Workplace-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Get all attendances
      tags:
      - Attendance
  /attendance/get_by_date:
    post:
      consumes:
      - application/json
      description: Retrieve all attendances within a specified date range for a given
        workplace
      parameters:
      - description: Workplace Token
        in: header
        name: X-Workplace-Token
        required: true
        type: string
      - description: Date Between
        in: body
        name: dateFrom
        required: true
        schema:
          $ref: '#/definitions/models.DateBetween'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                body:
                  items:
                    $ref: '#/definitions/models.AttendanceLaundry'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Get all attendances within a date range
      tags:
      - Attendance
  /attendance/get_by_employee/{employee_id}:
    get:
      consumes:
      - application/json
      description: Get Attendance by Employee ID
      parameters:
      - description: Workplace Token
        in: header
        name: X-Workplace-Token
        required: true
        type: string
      - description: ID of Employee
        in: path
        name: employee_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Get Attendance By Employee ID
      tags:
      - Attendance
  /attendance/update:
    put:
      consumes:
      - application/json
      description: Update Attendance by ID
      parameters:
      - description: Workplace Token
        in: header
        name: X-Workplace-Token
        required: true
        type: string
      - description: Employee body
        in: body
        name: attendanceUpdate
        required: true
        schema:
          $ref: '#/definitions/models.AttendanceUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: Update Attendance
      tags:
      - Attendance
...
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and the JWT token. Example: "Bearer
      eyJhbGciOiJIUz..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"